#1: refactor twitter consumers
